% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Categorical Proportional Function.R
\name{categorical_summary}
\alias{categorical_summary}
\title{Categorical Proportional Summary}
\usage{
categorical_summary(df, colstoremove = NULL)
}
\arguments{
\item{df}{The input data frame}

\item{colstoremove}{Any columns specified to be removed from the original data frame before running analysis}
}
\value{
The resulting proportions as printed data frames and ggplot visualizations of the resulting proportional data frames
}
\description{
Visualize and understand the proportion of each category within variables defined as character for a given data frame
}
