% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Categorical Proportional Function.R
\name{categoricalsummary}
\alias{categoricalsummary}
\title{Categorical Proportional Summary}
\usage{
categoricalsummary(df, colstoremove = NULL)
}
\arguments{
\item{df}{The input dataframe}

\item{colstoremove}{Any columns specified to be removed from the original dataframe before running analysis}
}
\description{
Visualize and understand the proportion of each category within variables defined as character for a given dataset
}
