% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KmeansAll.R
\name{kmeans_all}
\alias{kmeans_all}
\title{kmeans_all}
\usage{
kmeans_all(df)
}
\arguments{
\item{df}{The input data frame}
}
\description{
Apply the kmeans clustering algorithm to any data set and visualize commonalities between different categorical variables within each cluster. This function will automatically select the best number of clusters to apply to the data set by measuring the difference between each additional clsuter added to the analysis.
}
