% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NA Finder.R
\name{na_finder}
\alias{na_finder}
\title{na_finder}
\usage{
na_finder(df, colstoremove = NULL, Nulls = FALSE)
}
\arguments{
\item{df}{The input data frame}

\item{colstoremove}{Any columns specified to be removed from the original data frame before running analysis}

\item{Nulls}{If TRUE, will also look for strings equal to "NULL" and treat those rows as missing values, along with rows containing NA's}
}
\description{
Find the percentage of values that are NA/missing for each column in a data frame
}
