% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NA Finder.R
\name{nafinder}
\alias{nafinder}
\title{NA Finder Summary}
\usage{
nafinder(df, colstoremove = NULL)
}
\arguments{
\item{df}{The input dataframe}

\item{colstoremove}{Any columns specified to be removed from the original dataframe before running analysis}
}
\description{
Find the percentage of values that are NA/missing for each column in a dataframe
}
