% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ViolinAll.R
\name{violin_all}
\alias{violin_all}
\title{violin_all}
\usage{
violin_all(df, numericcolumn, colstoremove = NULL)
}
\arguments{
\item{df}{The input data frame}

\item{numericcolumn}{The numeric variable to be compared against each categorical variable}

\item{colstoremove}{Any columns specified to be removed from the original data frame before running analysis}
}
\description{
Create violin plots for each categorical variable in a data frame against a given numerical variable
}
